+++ package.json
@@ -10,6 +10,7 @@
   "dependencies": {
     "@material-ui/core": "^4.11.4",
     "@material-ui/icons": "^4.11.2",
+    "jquery": "^3.6.0",
     "next": "10.2",
     "next-auth": "^3.19.5",
     "react": "17.0.2",
