+++ public/js/custom.js
@@ -1,52 +1,3 @@
-jQuery("#owl-demo").owlCarousel({
-    loop: true,
-    nav: true,
-    navText: ["<div class='owl-prev'></div>", "<div class='owl-next'></div>"],
-    pagination: false,
-    items: 5,
-
-    responsiveClass: true,
-    responsive: {
-        0: {
-            items: 1,
-            nav: true,
-            loop: true
-        },
-        600: {
-            items: 2,
-            nav: false,
-            loop: true
-        },
-        1000: {
-            items: 3,
-            nav: true,
-            loop: true
-        },
-        1440: {
-            items: 4,
-            nav: true,
-            loop: true
-        },
-        1600: {
-            items: 5,
-            nav: true,
-            loop: true
-        }
-    }
-});
-
-jQuery("#owl-home-banner").owlCarousel({
-    autoplay: true,
-    loop: true,
-    navigation: false, // Show next and prev buttons
-    dots: false,
-    slideSpeed: 1000,
-    paginationSpeed: 400,
-    singleItem: true,
-    pagination: false,
-    items: 1,
-});
-
 jQuery("#owl-download-app").owlCarousel({
     autoplay: true,
     loop: true,
@@ -57,14 +8,6 @@ jQuery("#owl-download-app").owlCarousel({
     animateIn: 'fadeIn',
     pagination: false,
     items: 1,
-});
-jQuery("#scroller").simplyScroll({
-    customClass: 'vert',
-    orientation: 'vertical',
-    auto: true,
-    manualMode: 'loop',
-    speed: 1,
-    pauseOnTouch: true
 });
 
 jQuery(".ico-close").click(function(e) {
@@ -179,84 +122,6 @@ jQuery(document).ready(function(jQuery) {
         });
     });
 
-    jQuery('#owl-demo .slide').each(function () { 
-        var deadline = jQuery(this).data('date');
-        var number = jQuery(this).data('number');
-        var x = setInterval(function() { 
-            var now = new Date().getTime(); 
-            var timezoneOffset = new Date().getTimezoneOffset();
-            var t = (deadline + timezoneOffset*60*1000) - now;
-            var days = Math.floor(t / (1000 * 60 * 60 * 24));
-            var hours = Math.floor((t%(1000 * 60 * 60 * 24))/(1000 * 60 * 60));
-            var minutes = Math.floor((t % (1000 * 60 * 60)) / (1000 * 60));
-            var seconds = Math.floor((t % (1000 * 60)) / 1000);
-            jQuery("#owl-demo .caro_clock_"+number).html('<div class="itemCt"><div class="itemBg">'+ days + 'd ' + hours + ':' + minutes + ':' + seconds + '</div></div>');
-    
-            if (t < 0) {
-                clearInterval(x);
-                jQuery("#owl-demo .caro_clock_"+number).html('<div class="itemExpired"><div class="itemBg">EXPIRED</div></div>');
-            }
-        }, 1000);
-    });
-
-    jQuery('.exclusive-lotteries .track').each(function () {
-        var self = this;
-        function countdownDrawDate(deadline) {
-            var number = jQuery(self).data('number');
-            if (!deadline) {
-                deadline = jQuery(self).data('date');
-                if (number === 1) {
-                    deadline += (-180*60*1000);
-                }
-            } else {
-                var timezoneOffset = new Date().getTimezoneOffset();
-                if (number === 1) {
-                    deadline += (-180 - timezoneOffset)*60*1000;
-                } else {
-                    deadline += timezoneOffset*60*1000;
-                }
-            }
-    
-            var x = setInterval(function() { 
-                var now = new Date().getTime();
-                var t = deadline - now;
-                var days = Math.floor(t / (1000 * 60 * 60 * 24));
-                var hours = Math.floor((t%(1000 * 60 * 60 * 24))/(1000 * 60 * 60));
-                var minutes = Math.floor((t % (1000 * 60 * 60)) / (1000 * 60));
-                var seconds = Math.floor((t % (1000 * 60)) / 1000);
-    
-                var formattedHours = hours < 10 ? '0' + hours : hours;
-                var formattedMinutes = minutes < 10 ? '0' + minutes : minutes;
-                var formattedSeconds = seconds < 10 ? '0' + seconds : seconds;
-    
-                if (t < 0) {
-                    clearInterval(x);
-                    jQuery(".exclusive-lotteries .caro_clock_"+number).html('<div class="itemExpired"><div class="itemBg">EXPIRED</div></div>');
-    
-                    if (number === 1) {
-                        
-                    }
-                } else {
-                    jQuery(".exclusive-lotteries .caro_clock_"+number).html('<div class="itemCt"><div class="itemBg">Next Draw: '+ formattedHours + ':' + formattedMinutes + ':' + formattedSeconds + '</div></div>');
-                }
-            }, 1000);
-        }
-        countdownDrawDate();
-    });
-    
-    jQuery(function ($) {
-        jQuery(this).find("#middle_sec")
-        .click(function(){
-            jQuery(".show-sign-up").click();
-        });
-    });
-
-    jQuery(function ($) {
-        jQuery(".btn-signup").click(function(){
-            jQuery(".show-sign-up").click();
-        });
-    });
-    
     jQuery(document).on("click", ".custom-dropdown-menu-button", function() {
         var currentDataShow = jQuery(this).attr('data-show');
         jQuery(this).closest(".custom-dropdown-menu").find(".custom-dropdown-menu-items").toggle();
